requirements.txt
# LoAuth — Dependencies
# Core
flet>=0.25.0
pyotp>=2.9.0
cryptography>=42.0.0
argon2-cffi>=23.1.0
